[236-implement-redisset-data-structure-as-a-special-field-type] - extract base class to a sepearte file
